% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose_one_from_corr_snps.R
\name{choose_from_corr_snps}
\alias{choose_from_corr_snps}
\title{Choose quasi-independent SNPs from among correlated SNPs by using exposure association p-values and LD thresholding}
\usage{
choose_from_corr_snps(snp_info, ld_mat, ld_threshold = 0.5)
}
\arguments{
\item{snp_info}{a tibble with one row per SNP and columns "refsnp_id", "chr_name", "pos", "p_value".}

\item{ld_threshold}{a numeric value between 0 and 1. The threshold for pairwise LD values to consider two SNPs correlated.}

\item{ld_tibble}{a tibble with 3 columns: row_name, column_name, value. Typically the output from LDlinkR::LDmatrix and organize_ldlink_output}
}
\value{
a tibble with one row per quasi-independent SNP and columns "refsnp_id", "chr_name", "pos", "p_value"
}
\description{
Choose quasi-independent SNPs from among correlated SNPs by using exposure association p-values and LD thresholding
}
\details{
This function is used to choose a proxy SNP from among correlated SNPs by using exposure association p-values.
}
\examples{
snp_info <- tibble::tibble(refsnp_id = c("rs1", "rs2", "rs3"), chr_name = c("1", "1", "1"), 
     pos = c(100, 200, 300), p_value = c(0.01, 0.02, 0.03)) 
ld_threshold = 0.5
ld_tibble <- tibble::tibble(row_name = rep(c("rs1", "rs2", "rs3"), each = 3), 
 column_name = rep(c("rs1", "rs2", "rs3"), times = 3),
 value = c(1, 0.6, 0.2, 0.6, 1, 0.3, 0.2, 0.3, 1))
choose_one_from_corr_snps(snp_info, ld_tibble, ld_threshold)
 
}
