% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ldmatrix.R
\name{convert_df_to_matrix}
\alias{convert_df_to_matrix}
\title{Convert ld_df to a R matrix}
\usage{
convert_df_to_matrix(ld_df)
}
\arguments{
\item{ld_df}{typically the output from get_ld_df}
}
\value{
a square R matrix containing LD values
}
\description{
Convert ld_df to a R matrix
}
