% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_correlated_snps.R
\name{get_correlated_snps}
\alias{get_correlated_snps}
\title{Get SNPs in high LD with each lead SNP}
\usage{
get_correlated_snps(
  snps,
  genome_build,
  ld_token,
  pop = "CEU",
  r2_threshold = 0.8
)
}
\arguments{
\item{snps}{a character vector of SNPs - either rsids or chromosome:position}

\item{genome_build}{a character string of the genome build}

\item{ld_token}{a character string of the LDlinkR token}

\item{pop}{a character string of the population to use for LD calculation}

\item{r2_threshold}{a numeric value of the r2 threshold to use in filtering correlated SNPs}
}
\value{
a tibble with one row per correlated SNP
}
\description{
Get SNPs in high LD with each lead SNP
}
\details{
queries LDlink API once per lead SNP
}
