% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ldmatrix.R
\name{get_ld_tib}
\alias{get_ld_tib}
\title{Get LD values for a collection of SNPs on one chromosome}
\usage{
get_ld_tib(snps, genome_build, token, pop = "CEU")
}
\arguments{
\item{snps}{a character vector of SNPs}

\item{host}{a character string of the host to use for the LDlinkR query}
}
\value{
a tibble with 3 columns: row_name, column_name, value
}
\description{
Get LD values for a collection of SNPs on one chromosome
}
\details{
This function is used to get LD values for a collection of SNPs.
}
