% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pos_from_rs_id.R
\name{get_pos_from_rs_id}
\alias{get_pos_from_rs_id}
\title{Get GRCh38 position AND GRCh37 for lead SNP}
\usage{
get_pos_from_rs_id(rs_id)
}
\arguments{
\item{rs_id}{a character vector of length one, ie, one SNP rs id value for one lead SNP}
}
\value{
a tibble with one row and columns for both GRCh37 and GRCh38 positions
}
\description{
Get GRCh38 position AND GRCh37 for lead SNP
}
\details{
queries biomaRt API twice per SNP
}
