% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rsid_from_chr_pos.R
\name{get_rsid_from_chr_pos}
\alias{get_rsid_from_chr_pos}
\title{Get rs_id from chromosome and position}
\usage{
get_rsid_from_chr_pos(chr, pos, host = "https://grch37.ensembl.org")
}
\arguments{
\item{chr}{a character vector of length 1- the chromosome}

\item{pos}{a numeric vector of snp positions on the chromosome chr}

\item{host}{a character string of the host to use for the biomaRt query}
}
\value{
a tibble with one row per SNP
}
\description{
Get rs_id from chromosome and position
}
\details{
queries biomaRt API once per SNP
}
