% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_snps_near_lead.R
\name{get_snps_near_lead}
\alias{get_snps_near_lead}
\title{Get SNPs near a lead SNP and in a tibble of SNPs, like those from GWAS summary statistics text files}
\usage{
get_snps_near_lead(
  lead_snp_pos,
  lead_snp_chrom,
  dat,
  dat_chrom_var,
  dat_pos_var,
  dist = 1e+05
)
}
\arguments{
\item{lead_snp_pos}{an integer vector of length one}

\item{lead_snp_chrom}{an integer vector of length one to indicate which numeric chromosome SNP is on}

\item{dat}{a tibble of SNPs, like those from GWAS summary statistics text files}

\item{dat_chrom_var}{the name of the variable in dat that indicates which chromosome the SNP is on}

\item{dat_pos_var}{the name of the variable in dat that indicates the position of the SNP}

\item{dist}{the distance in base pairs to search for SNPs near the lead SNP}
}
\value{
a tibble of SNPs near the lead SNP
}
\description{
Get SNPs near a lead SNP and in a tibble of SNPs, like those from GWAS summary statistics text files
}
