% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organize_ldlink_output.R
\name{organize_ldlink_output}
\alias{organize_ldlink_output}
\title{Organize outputted LD matrix (a data.frame) from LDlink into 3-column tibble & discard NAs}
\usage{
organize_ldlink_output(ldlink_output)
}
\arguments{
\item{ldlink_output}{a data.frame from LDlinkR::LDmatrix}
}
\value{
a tibble with 3 columns: row_name, column_name, value
}
\description{
Organize outputted LD matrix (a data.frame) from LDlink into 3-column tibble & discard NAs
}
