% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_correlated_alleles.R
\name{parse_correlated_alleles}
\alias{parse_correlated_alleles}
\title{Parse correlated alleles}
\usage{
parse_correlated_alleles(corr_alleles)
}
\arguments{
\item{corr_alleles}{a character vector of correlated alleles, typically outputted from LDlinkR::LDmatrix}
}
\description{
Parse correlated alleles
}
