% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_alleles.R
\name{parse_correlated_alleles}
\alias{parse_correlated_alleles}
\title{Parse ldlink output for proxy SNPs' alleles in column called Correlated_Alleles}
\usage{
parse_correlated_alleles(allele_string)
}
\arguments{
\item{allele_string}{a character string of alleles from the Correlated_Alleles column with form like A=A,G=C}
}
\value{
a tibble with two rows per proxy SNP
}
\description{
Parse ldlink output for proxy SNPs' alleles in column called Correlated_Alleles
}
\details{
lead SNP's alleles are always on the left side of the equals sign in the Correlated_Alleles column
}
\examples{
a_string <- "A=A,G=C"
parse_correlated_alleles(a_string)
}
